pull out snapshots into separate file
