+++ tests/contrib/django/test_django.py
@@ -15,7 +15,7 @@ from ddtrace.ext.priority import USER_KEEP
 from ddtrace.propagation.http import HTTP_HEADER_TRACE_ID, HTTP_HEADER_PARENT_ID, HTTP_HEADER_SAMPLING_PRIORITY
 from ddtrace.propagation.utils import get_wsgi_header
 
-from tests import override_config, override_global_config, override_http_config, assert_dict_issuperset, snapshot
+from tests import override_config, override_global_config, override_http_config, assert_dict_issuperset
 from tests.opentracer.utils import init_tracer
 
 pytestmark = pytest.mark.skipif("TEST_DATADOG_DJANGO_MIGRATION" in os.environ, reason="test only without migration")
@@ -500,27 +500,6 @@ def test_middleware_trace_function_based_view(client, test_spans):
         assert span.resource == "GET tests.contrib.django.views.function_view"
 
 
-@pytest.mark.skipif(django.VERSION > (1, 12), reason="")
-@snapshot()
-def test_middleware_trace_callable_view_111x(client):
-    # ensures that the internals are properly traced when using callable views
-    assert client.get("/feed-view/").status_code == 200
-
-
-@pytest.mark.skipif(not (django.VERSION > (1, 12) and django.VERSION < (2, 2)), reason="")
-@snapshot()
-def test_middleware_trace_callable_view_21x(client):
-    # ensures that the internals are properly traced when using callable views
-    assert client.get("/feed-view/").status_code == 200
-
-
-@pytest.mark.skipif(django.VERSION < (2, 2), reason="")
-@snapshot()
-def test_middleware_trace_callable_view(client):
-    # ensures that the internals are properly traced when using callable views
-    assert client.get("/feed-view/").status_code == 200
-
-
 def test_middleware_trace_errors(client, test_spans):
     # ensures that the internals are properly traced
     assert client.get("/fail-view/").status_code == 403
@@ -545,27 +524,6 @@ def test_middleware_trace_staticmethod(client, test_spans):
         assert span.resource == "GET ^static-method-view/$"
     else:
         assert span.resource == "GET tests.contrib.django.views.StaticMethodView"
-
-
-@pytest.mark.skipif(django.VERSION > (1, 12), reason="")
-@snapshot()
-def test_middleware_trace_partial_based_view_111x(client):
-    # ensures that the internals are properly traced when using a function views
-    assert client.get("/partial-view/").status_code == 200
-
-
-@pytest.mark.skipif(not (django.VERSION > (1, 12) and django.VERSION < (2, 2)), reason="")
-@snapshot()
-def test_middleware_trace_partial_based_view_21x(client):
-    # ensures that the internals are properly traced when using a function views
-    assert client.get("/partial-view/").status_code == 200
-
-
-@pytest.mark.skipif(django.VERSION < (2, 2), reason="")
-@snapshot()
-def test_middleware_trace_partial_based_view(client):
-    # ensures that the internals are properly traced when using a function views
-    assert client.get("/partial-view/").status_code == 200
 
 
 def test_simple_view_get(client, test_spans):
@@ -1366,26 +1324,6 @@ def test_urlpatterns_path(client, test_spans):
 
     # Ensure the view was traced
     assert len(list(test_spans.filter_spans(name="django.view"))) == 1
-
-
-@pytest.mark.skipif(not (django.VERSION > (2, 0) and django.VERSION < (2, 2)), reason="")
-@snapshot()
-def test_urlpatterns_include_21x(client):
-    """
-    When a view is specified using `django.urls.include`
-        The view is traced
-    """
-    assert client.get("/include/test/").status_code == 200
-
-
-@pytest.mark.skipif(django.VERSION < (2, 2), reason="")
-@snapshot()
-def test_urlpatterns_include(client):
-    """
-    When a view is specified using `django.urls.include`
-        The view is traced
-    """
-    assert client.get("/include/test/").status_code == 200
 
 
 @pytest.mark.skipif(django.VERSION < (2, 0, 0), reason="repath only exists in >=2.0.0")
